+++ libs/service-portal/core/src/index.ts
@@ -4,6 +4,7 @@ export * from './lib/navigation/paths'
 export * from './components/NavigationScreen/NavigationScreen'
 export * from './components/ActionCard/ActionCard'
 export * from './components/ActionCard/ActionCardLoader'
+export * from './components/EducationCard/EducationCard'
 export * from './components/ActionMenu/ActionMenu'
 export * from './components/EventCard/EventCard'
 export * from './components/Menu/Menu'
