+++ libs/service-portal/education-degree/src/screens/EducationDegree/components/DegreeCards/index.ts
@@ -1,0 +1,1 @@
+export { default as DegreeCards } from './DegreeCards'
