--- libs/service-portal/education-degree/src/screens/EducationDegree/components/DegreeTable/index.ts
@@ -1,1 +1,0 @@
-export { default as DegreeTable } from './DegreeTable'
