--- libs/service-portal/education-license/src/screens/EducationLicense/components/LicenseTable/index.ts
@@ -1,1 +1,0 @@
-export { default as LicenseTable } from './LicenseTable'
